fix compile warnings, db error
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@187 26674162-90ff-0310-a85b-b64107c8dee9

+++ src/game/Chat.cpp
@@ -344,7 +344,7 @@ void ChatHandler::FillMessageData( WorldPacket *data, WorldSession* session, uin
     else if (type == CHAT_MSG_WHISPER_INFORM)
     {
         // Convert ChannelName back to the to Players GUID
-        guid = uint64(channelName);               //session ? session->GetPlayer()->GetGUID() : 0; // FIXME: may be receiver?
+        guid = uint32(channelName);               //session ? session->GetPlayer()->GetGUID() : 0; // FIXME: may be receiver?
     }
 
     *data << guid;
